Move not shared to client
