+++ packages/react-dom/src/client/CSSShorthandProperty.js
