+++ packages/react-dom/src/client/createMicrosoftUnsafeLocalFunction.js
